# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: CipherQuest API
  description: |
    A comprehensive cybersecurity learning platform API that provides authentication, 
    module management, CTF challenges, leaderboards, and user progress tracking.
    
    ## Features
    - JWT-based authentication with OAuth support (Google, GitHub)
    - Learning modules with progress tracking
    - CTF challenges with flag submission
    - Real-time leaderboards
    - User profile and progress management
    - Admin dashboard and content management
    
    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```
    
    ## Rate Limiting
    - Auth endpoints: 5-10 requests per minute
    - Challenge submission: 20 requests per minute
    - Other endpoints: 30-60 requests per minute
  version: 1.0.0
  contact:
    name: CipherQuest Support
    email: support@cipherquest.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cipherquest.com
    description: Production server
  - url: https://cipherquest-api.herokuapp.com
    description: Heroku deployment
  - url: http://localhost:5000
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /api/challenges:
    get:
      tags:
        - Challenges
      summary: Get all challenges
      description: Retrieve all CTF challenges with optional filtering
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by challenge category
          schema:
            type: string
            example: "web"
        - name: difficulty
          in: query
          description: Filter by difficulty level
          schema:
            type: string
            enum: [easy, medium, hard]
            example: "easy"
        - name: module_id
          in: query
          description: Filter by associated module
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Number of challenges to return
          schema:
            type: integer
            default: 50
            maximum: 100
            example: 20
        - name: offset
          in: query
          description: Number of challenges to skip
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: List of challenges
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenges:
                    type: array
                    items:
                      $ref: '#/components/schemas/Challenge'
                  total:
                    type: integer
                    example: 50
                  limit:
                    type: integer
                    example: 20
                  offset:
                    type: integer
                    example: 0
  /api/challenges/{challenge_id}:
    get:
      tags:
        - Challenges
      summary: Get challenge details
      description: Retrieve detailed information about a specific challenge
      security:
        - BearerAuth: []
      parameters:
        - name: challenge_id
          in: path
          required: true
          description: Challenge ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Challenge details
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    $ref: '#/components/schemas/Challenge'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/challenges/{challenge_id}/submit:
    post:
      tags:
        - Challenges
      summary: Submit flag
      description: Submit a flag for a challenge
      security:
        - BearerAuth: []
      parameters:
        - name: challenge_id
          in: path
          required: true
          description: Challenge ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - flag
              properties:
                flag:
                  type: string
                  description: The flag to submit
                  example: "flag{this_is_a_test_flag}"
      responses:
        '200':
          description: Flag submission result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Flag is correct!"
                  points_earned:
                    type: integer
                    example: 50
                  progress:
                    $ref: '#/components/schemas/UserProgress'
                  attempts:
                    type: integer
                    example: 3
        '400':
          description: Invalid flag format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/challenges/{challenge_id}/hint:
    get:
      tags:
        - Challenges
      summary: Get challenge hint
      description: Retrieve a hint for a challenge
      security:
        - BearerAuth: []
      parameters:
        - name: challenge_id
          in: path
          required: true
          description: Challenge ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Challenge hint
          content:
            application/json:
              schema:
                type: object
                properties:
                  hint:
                    type: string
                    example: "Check the source code of the page"
                  hint_cost:
                    type: integer
                    example: 10
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Challenge:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "SQL Injection Basics"
        description:
          type: string
          example: "Learn about SQL injection vulnerabilities"
        category:
          type: string
          example: "web"
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "easy"
        points:
          type: integer
          example: 50
        module_id:
          type: integer
          example: 1
        is_active:
          type: boolean
          example: true
        user_progress:
          $ref: '#/components/schemas/UserProgress'
          nullable: true
    UserProgress:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        module_id:
          type: integer
          nullable: true
          example: 1
        challenge_id:
          type: integer
          nullable: true
          example: 1
        completed:
          type: boolean
          example: true
        completed_at:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-20T15:30:00Z"
        attempts:
          type: integer
          example: 3
        time_spent:
          type: integer
          description: Time spent in seconds
          example: 1800
        score:
          type: integer
          example: 100
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid credentials"
        message:
          type: string
          example: "The provided credentials are incorrect"
        code:
          type: string
          example: "AUTH_001"
        details:
          type: object
          additionalProperties: true
...